refactor: move deprecated_command to separate file and improve warning messaging
